First version of testing. Drop Python3.7

+++ src/pyBL/dataclasses/storm.py
@@ -6,8 +6,8 @@ from datetime import timedelta
 
 @dataclass(frozen=True)
 class Cell:
-    startTime: int
-    endTime: int
+    startTime: float
+    endTime: float
     intensity: float
     timescale: timedelta = timedelta(hours=1)
